CMake as build system, Google Test as unit test framework.
The building system is almost completed, each practice is a independent target,
and every source will be linted by cpplint.py which can detect the violation of
Google C++ Style Guide.
Google Test is ready to use, if there is main function in source code of
practice, then the main function of Google Test will not be executed. So if you
want to use Google Test, you must ensure that you haven't write main function in
your source code.